# docs/swagger.json
{"openapi":"3.0.0","paths":{"/api/generic-adapter/v1":{"get":{"operationId":"AppController_find","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"default":10,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PageDto"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/GenericInterface"}}}}]}}}}}},"post":{"operationId":"AppController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGenericDataDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}}}}},"/api/generic-adapter/v1/{id}":{"get":{"operationId":"AppController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}}}},"put":{"operationId":"AppController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGenericDataDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericInterface"}}}}}},"delete":{"operationId":"AppController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Generic Adapter ","description":"Adapter of Generic backend repo","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"PageMetaDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","take","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PageDto":{"type":"object","properties":{"data":{"type":"array","items":{"type":"array"}},"meta":{"$ref":"#/components/schemas/PageMetaDto"}},"required":["data","meta"]},"GenericInterface":{"type":"object","properties":{"id":{"type":"number"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","description","createdAt","updatedAt"]},"CreateGenericDataDto":{"type":"object","properties":{"description":{"type":"string"}},"required":["description"]},"UpdateGenericDataDto":{"type":"object","properties":{"description":{"type":"string"}}}}}}